+++ src/data/stickers.json
